Move ConcreteBaseAudioContext to separate file
and define some glue functions. In the old situation, attributes of the
concrete implementation would be accessed directly
